Moved start message printer into HttpStub
